renaming documentation, directory/packages, tests
