+++ internal/service/team/resource_team.go
@@ -1,4 +1,4 @@
-package teams
+package team
 
 import (
 	"context"
